# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaaS Multi-Niche API",
        "description": "API RESTful pour système de réservation multi-niche (paragliding, surfing, diving, etc.)",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server (Développement)"
        },
        {
            "url": "https://api.example.com",
            "description": "API Server (Production)"
        }
    ],
    "paths": {
...
        "/api/v1/options": {
            "get": {
                "tags": [
                    "Options"
                ],
                "summary": "Liste des options",
                "description": "Retourne la liste des options disponibles (photos, vidéos, etc.)",
                "operationId": "listOptions",
                "responses": {
                    "200": {
                        "description": "Liste des options",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "organization": []
                    }
                ]
            }
        },
        "/api/v1/admin/options": {
            "post": {
                "tags": [
                    "Options"
                ],
                "summary": "Créer une option (Admin)",
                "description": "Crée une nouvelle option (photo, vidéo, souvenir, etc.)",
                "operationId": "createOption",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "price_type",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Photos HD"
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 25
                                    },
                                    "price_type": {
                                        "type": "string",
                                        "enum": [
                                            "fixed",
                                            "per_participant"
                                        ],
                                        "example": "fixed"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "photo",
                                            "video",
                                            "souvenir",
                                            "insurance",
                                            "transport",
                                            "other"
                                        ],
                                        "example": "photo"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Option créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Option créée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "organization": []
                    }
                ]
            }
        },
        "/api/v1/admin/options/{id}": {
            "put": {
                "tags": [
                    "Options"
                ],
                "summary": "Modifier une option (Admin)",
                "description": "Met à jour une option existante",
                "operationId": "updateOption",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Photos HD"
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 25
                                    },
                                    "price_type": {
                                        "type": "string",
                                        "enum": [
                                            "fixed",
                                            "per_participant"
                                        ],
                                        "example": "fixed"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "photo",
                                            "video",
                                            "souvenir",
                                            "insurance",
                                            "transport",
                                            "other"
                                        ],
                                        "example": "photo"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Option mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Option mise à jour avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Option non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "organization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Options"
                ],
                "summary": "Supprimer une option (Admin)",
                "description": "Supprime une option du système",
                "operationId": "deleteOption",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Option supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Option supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Option non trouvée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "organization": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token. Obtenez le token via /api/v1/auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "organization": {
                "type": "apiKey",
                "description": "ID de l'organisation (tenant). Requis pour l'isolation multi-tenant.",
                "name": "X-Organization-ID",
                "in": "header"
            }
        }
    },
...
}